+++ source/PCM_API.h
@@ -71,7 +71,7 @@ void PCM_set_attenuation_level_left(uint8_t value);
  *                  15 - 255		(-120,0) - dB 0 dB
  *                  0 - 14          mute
  */
-uint8_t PCM_get_attenuation_level_left(void);
+uint16_t PCM_get_attenuation_level_left(void);
 
 /**
  * @brief           set digital attenuation level, right side
@@ -93,7 +93,7 @@ void PCM_set_attenuation_level_right(uint8_t value);
  *                  15 - 255		(-120,0) - dB 0 dB
  *                  0 - 14          mute
  */
-uint8_t PCM_get_attenuation_level_right(void);
+uint16_t PCM_get_attenuation_level_right(void);
 
 /** ATS: Attenuation Rate Select 
     read and write */
@@ -111,7 +111,7 @@ void PCM_set_soft_mute(uint8_t value);
  * 
  * @return          disabled/enabeld
  */
-uint8_t PCM_get_soft_mute(void);
+uint16_t PCM_get_soft_mute(void);
 
 /**
  * @brief           set ATLD: attenuation load control
@@ -133,7 +133,7 @@ void PCM_set_attenuation_control(uint8_t value);
  * 
  * @return          disabled/enabeld
  */
-uint8_t PCM_get_attenuation_control(void);
+uint16_t PCM_get_attenuation_control(void);
 
 /**
  * @brief           set audio interface data format
@@ -223,7 +223,7 @@ void PCM_set_zero_detect_mute(uint8_t value);
  * 
  * @return          disabled/enabeld
  */
-uint8_t PCM_get_zero_detect_mute(void);
+uint16_t PCM_get_zero_detect_mute(void);
 
 /**
  * @brief           system reset control 
@@ -296,14 +296,14 @@ PCM_data_channel_t PCM_get_data_channel(void);
  * 
  * @return          not zero = 0, zero detect = 1
  */
-uint8_t PCM_get_left_zero_detection(void);
+uint16_t PCM_get_left_zero_detection(void);
 
 /**
  * @brief           zero detect in output channel, right
  * 
  * @return          not zero = 0, zero detect = 1
  */
-uint8_t PCM_get_rigth_zero_detection(void);
+uint16_t PCM_get_rigth_zero_detection(void);
 
 /** ID: Device ID 
     read only */
